feat(region): Move test to actions-test
