+++ src/store/highlight/__tests__/reducer-test.ts
@@ -2,7 +2,7 @@ import reducer from '../reducer';
 import state from '../__mocks__/highlightState';
 import { Annotation, NewAnnotation } from '../../../@types';
 import { createAnnotationAction } from '../../annotations';
-import { mockContainerRect, mockDOMRect, mockRange } from '../__mocks__/data';
+import { mockContainerRect, mockRange } from '../__mocks__/data';
 import { resetCreatorAction } from '../../creator';
 import { setIsPromotingAction, setSelectionAction } from '../actions';
 
@@ -26,28 +26,6 @@ describe('store/highlight/reducer', () => {
             const newState = reducer(state, setSelectionAction(payload));
 
             expect(newState.selection).toEqual({ ...state.selection, location: 2 });
-        });
-
-        test('should set selection in IE/Edge', () => {
-            jest.spyOn(document, 'createNodeIterator').mockReturnValueOnce(({
-                nextNode: jest.fn().mockReturnValueOnce(mockRange.startContainer),
-            } as unknown) as NodeIterator);
-            jest.spyOn(document, 'createRange').mockReturnValueOnce({
-                ...new Range(),
-                getBoundingClientRect: jest.fn().mockReturnValueOnce(mockDOMRect),
-                selectNodeContents: jest.fn(),
-                setEnd: jest.fn(),
-                setStart: jest.fn(),
-            });
-            const range = {
-                ...mockRange,
-                getClientRects: () => ({ length: 0 } as DOMRectList),
-            };
-
-            const payload = { containerRect: mockContainerRect, location: 1, range };
-            const newState = reducer(state, setSelectionAction(payload));
-
-            expect(newState.selection).toEqual(state.selection);
         });
     });
